refactor: Move Cats instances to their own file
